+++ Changelog.md
@@ -9,3 +9,4 @@
 - Rename env var used to save request examples from `EXPLICIT_PERSIST_EXAMPLES`
   to `UPDATE_REQUEST_EXAMPLES`.
 - Add `company_logo_url(url)` to documentation builder.
+- Drop `parse: true` option from the following types: `:agreement`.